refactor: 迁移 session_id 从 uint64_t 到 uint32_t
- 修改协议常量 P2P_SESS_ID_PSZ 从 8字节 到 4字节
- 更新所有结构体 session_id 字段类型 (session_t, compact_session_t, relay_session_context_t, signal_compact_session_t)
- 替换所有网络序列化函数: nwrite_ll/nread_ll → nwrite_l/nread_l
- 修复哈希表键长度不一致 bug (HASH_ADD 使用 4字节，HASH_FIND 之前使用 8字节)
- 更新所有格式化字符串: PRIu64 → PRIu32
- 更新函数参数和局部变量类型
- 修改数据包格式注释和协议文档

影响:
- 协议破坏性变更，需客户端服务器同步升级
- SYNC0_ACK 包大小从 13字节 减少到 9字节
- session_id 碰撞空间从 2^64 降至 2^32 (仍使用加密安全 RNG)

验证:
- 编译通过，无新增错误或警告
- p2p_server 和 p2p_ping 构建成功
- 所有测试目标编译通过

+++ include/p2pp.h
@@ -23,7 +23,7 @@
 #define P2P_PEER_ID_MAX     32
 
 /* Session ID 字节长度 */
-#define P2P_SESS_ID_PSZ     (sizeof(uint64_t))
+#define P2P_SESS_ID_PSZ     (sizeof(uint32_t))
 
 /* ============================================================================
  * 可交换的候选地址定义
